+++ crates/siril-sys/src/lib.rs
@@ -11,5 +11,3 @@ pub use siril::Builder;
 pub use siril::Siril;
 pub use siril::find_siril_cli;
 pub use types::*;
-
-// TODO: Siril macro for easy one off jobs