+++ crates/siril-sys/src/siril_ext.rs
@@ -1,5 +1,6 @@
 //! Prelude file for all Siril Extensions from Commands
 //! import with `use siril_sys::siril_ext::*;`
 
+pub use crate::commands::cd::CdExt;
 pub use crate::commands::load::LoadExt;
 pub use crate::commands::pwd::PwdExt;